+++ CHANGELOG.md
@@ -67,6 +67,8 @@
 - New utility functions:
   1. `len(Array|Bytes|Script) -> Number`
   1. `typeof(Value) -> String`
+  1. `int(Float) -> Integer`
+  1. `float(Int) -> Float`
   1. `keys(Array) -> Array<Number>`
   2. `first(Array) -> Any`
   3. `last(Array) -> Any`